# apps/api/src/swagger/api-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_getUsers",
        "parameters": [
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUsersResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "UsersController_deleteBulkUsers",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteBulkUsersBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteBulkUsersResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/user": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserByIdResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/user-details": {
      "get": {
        "operationId": "UsersController_getUserDetails",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserDetailsResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "UsersController_upsertUserDetails",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertUserDetailsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpsertUserDetailsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/admin/user": {
      "patch": {
        "operationId": "UsersController_adminUpdateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUpdateUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/change-password": {
      "patch": {
        "operationId": "UsersController_changePassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/create": {
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files": {
      "post": {
        "operationId": "S3Controller_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "resource": {
                    "type": "string",
                    "description": "Optional resource type"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "S3Controller_deleteFile",
        "parameters": [
          {
            "name": "fileKey",
            "required": true,
            "in": "query",
            "description": "Key of the file to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully"
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Guidebook API",
    "description": "Example usage of Swagger with Typebox",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "GetUsersResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "archived": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "createdAt",
                "updatedAt",
                "email",
                "firstName",
                "lastName",
                "role",
                "archived"
              ]
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "totalItems": {
                "type": "number"
              },
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              }
            },
            "required": [
              "totalItems",
              "page",
              "perPage"
            ]
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
      "GetUserByIdResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "createdAt": {
                "type": "string"
              },
              "updatedAt": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "role": {
                "type": "string"
              },
              "archived": {
                "type": "boolean"
              }
            },
            "required": [
              "id",
              "createdAt",
              "updatedAt",
              "email",
              "firstName",
              "lastName",
              "role",
              "archived"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "GetUserDetailsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "format": "uuid",
                "type": "string"
              },
              "description": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "contactEmail": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "contactPhone": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "jobTitle": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "id",
              "description",
              "contactEmail",
              "contactPhone",
              "jobTitle"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateUserBody": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "const": "admin",
                "type": "string"
              },
              {
                "const": "student",
                "type": "string"
              },
              {
                "const": "teacher",
                "type": "string"
              }
            ]
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
      "UpdateUserResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "createdAt": {
                "type": "string"
              },
              "updatedAt": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "role": {
                "type": "string"
              },
              "archived": {
                "type": "boolean"
              }
            },
            "required": [
              "id",
              "createdAt",
              "updatedAt",
              "email",
              "firstName",
              "lastName",
              "role",
              "archived"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "UpsertUserDetailsBody": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "contactEmail": {
            "format": "email",
            "type": "string"
          },
          "contactPhoneNumber": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          }
        }
      },
      "UpsertUserDetailsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "format": "uuid",
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "message"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "AdminUpdateUserBody": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "const": "admin",
                "type": "string"
              },
              {
                "const": "student",
                "type": "string"
              },
              {
                "const": "teacher",
                "type": "string"
              }
            ]
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
      "AdminUpdateUserResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "createdAt": {
                "type": "string"
              },
              "updatedAt": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "role": {
                "type": "string"
              },
              "archived": {
                "type": "boolean"
              }
            },
            "required": [
              "id",
              "createdAt",
              "updatedAt",
              "email",
              "firstName",
              "lastName",
              "role",
              "archived"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "ChangePasswordBody": {
        "type": "object",
        "properties": {
          "newPassword": {
            "minLength": 8,
            "maxLength": 64,
            "type": "string"
          },
          "oldPassword": {
            "minLength": 8,
            "maxLength": 64,
            "type": "string"
          }
        },
        "required": [
          "newPassword",
          "oldPassword"
        ]
      },
      "ChangePasswordResponse": {
        "type": "null"
      },
      "DeleteUserResponse": {
        "type": "null"
      },
      "DeleteBulkUsersBody": {
        "type": "object",
        "properties": {
          "userIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userIds"
        ]
      },
      "DeleteBulkUsersResponse": {
        "type": "null"
      },
      "CreateUserBody": {
        "type": "object",
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "firstName": {
            "minLength": 1,
            "maxLength": 64,
            "type": "string"
          },
          "lastName": {
            "minLength": 1,
            "maxLength": 64,
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "const": "admin",
                "type": "string"
              },
              {
                "const": "student",
                "type": "string"
              },
              {
                "const": "teacher",
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "role"
        ]
      },
      "CreateUserResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "format": "uuid",
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "message"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
...
      "FileUploadResponse": {
        "type": "object",
        "properties": {
          "fileKey": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          }
        },
        "required": [
          "fileKey",
          "fileUrl"
        ]
      },
...
    }
  }
}